Attempt rollback function on uncaught throwables
Prior to this change, the rollback function on uncaught throwables
raised by failed shared ref db updates would not attempt to roll back
the local update. This would sometimes cause split brain issues.

This change conditionally executes the rollback in a finally block when
the shared ref db update fails. The rollback may fail if the jvm is
already in an unrecoverable broken state.

When the rollback fails we now also return an IO_FAILURE result instead
of a LOCK_FAILURE to indicate that our system is now in an uncertain
state.

Bug: Issue 297890487
Change-Id: I9be963590743065499e1365f1e92dd02b1fffcb0