Handle successful and failed rollbacks aptly
Previously we didn't handle rollback failures at all. This meant the
logs didn't output any additional information when the local ref was
out-of-sync with the global refdb.

This change adds logging and amends the return result.

Note that because we are in a finally block we cannot return directly or
throw, or we'll clobber any existing throwable.

Bug: Issue 297890487
Change-Id: I74fe9d938f7b315c8dde0efac2abcf34b7ded9b2