Attempt rollback function on uncaught throwables
Prior to this change, when failed shared ref db updates would raise
uncaught throwables, the rollback function would not be executed. This
could cause split brain issues.

This change conditionally executes the rollback in a finally block when
the shared ref db update fails. Note that the rollback may fail if the
JVM is already in an broken state.

Bug: Issue 297890487
Change-Id: I9be963590743065499e1365f1e92dd02b1fffcb0