[RFC] Improve handling of globali refdb update failures
For a BatchRefUpdate for two or more ref updates, where all the ref
updates were successfully applied locally, but at least one failed in
global refdb, the behaviour of the system was to rollback both ref
updates locally. This results in the successfully applied changes
in global refdb to be ahead of the local changes. Moreover, these
refs will now be only in global refdb and nowhere else!

What the system should do instead, is not rollback the local changes,
log exactly what happened and let the gerrit admin decide how to
resolve the split brain.

Stop rolling back local changes when updates in global refdb fail.
Increment the split brain metric, and mark the status of the ref
updates as lock failure.

From the code it looks like the `updateSharedRefDb` can throw a
couple of different exceptions, with different semantics. One is
the `GlobalRefDbSystemError` which will be thrown when `compareAndPut`
throws an exception. The other is `SharedDbSplitBrainException` which
will be thrown when `compareAndPut` does not succeed. Regardless,
I think the result is the same (ie the global refdb failed to get
updated, but locally the ref updates have been applied) and as such
we shouldn't treat each exception in a different way.

Regarding logging, both exceptions will log exactly which project
and ref was affected, so there is no point logging this again after
we catch the exception. However, I think there should be a follow-up
change to include the object id value in the `GlobalRefDbSystemError`
exception message.

Bug: Issue 300831672
Change-Id: I421183997dbeb80f0f8da7e84a7b8046315364f2