Throw LockFailureException-based exception when failing to get lock
Gerrit has retries in cases that a ref-update fails with a
LockFailureException. The global-refdb didn't use that Exception, but
threw a RuntimeException (GlobalRefDbLockException). Thus, in case the
lock in the global refdb could not be acquired, the ref-update was
aborted after the first try. This could lead to a lot of failed updates
like reviews, affecting especially users on very active repositories.

This change changes the Exception hierarchy, so that SharedLockException
inherits from LockFailureException, which also inherits from IOException
and is thus not breaking the API. The GlobalRefDbException now inherits
from SharedLockException and is thus not a RuntimeException anymore and
will thus be handed down to the RetryHelper.

The GlobalRefDbException is now only an alias for SharedRefDbException,
but was kept to not break the API.

Change-Id: I3518ba9187761f8475233867ed8a3212713c9079