Retry on lock failure during change update
Gerrit does not retry if posting a review fails with a lock failure. Thus,
in case the lock in the global refdb could not be acquired, the ref-update
was aborted after the first try. This could lead to a lot of failed updates
like reviews, affecting especially users on very active repositories.

This change changes the Exception hierarchy, so that SharedLockException
inherits from LockFailureException, which also inherits from IOException
and is thus not breaking the API. The GlobalRefDbException now inherits
from SharedLockException and is thus not a RuntimeException anymore.

The GlobalRefDbException is now only an alias for SharedRefDbException,
but was kept to not break the API.

Further an ExceptionHook has been added that will cause Gerrit to retry, if
a LockFailureException is being thrown. This ExceptionHook has to be bound
by every plugin that wants to make use of the Global RefDB.

Change-Id: I3518ba9187761f8475233867ed8a3212713c9079