Retry on lock failure during change update
Gerrit does not retry if posting a review fails with a lock failure
thrown by the global refdb implementation. Thus, in case the lock in
the global refdb could not be acquired, the ref-update was aborted
after the first try. This could lead to a lot of failed updates like
reviews, affecting especially users on very active repositories.

An ExceptionHook has been added that will cause Gerrit to retry,
if a LockFailureException is being thrown. This ExceptionHook has to be
bound by every plugin that wants to make use of the Global RefDB.

Bug: Issue 334278785
Change-Id: I3518ba9187761f8475233867ed8a3212713c9079