Enforce draftComments when their streaming is enabled
Change I02e17592b introduced the possibility to make draft comment
events publishing configurable.

As a consequence of this, replication mechanisms such as the replication
and pull-replication plugins might replicate draft-comments refs on
different Gerrit nodes.

To avoid split-brain, they have to be treated as any other ref and thus
their global status needs to be enforced at a global refdb level.

Account for the `event.stream-events.enableDraftCommentEvents` option
when evaluating the default policy enforcement for the `draft-comments`
ref-space and ignore them only when their streaming is disabled.

Bug: Issue 367623925
Change-Id: I269ae1bc928e8e813740ebbed925b94472dc0063