Add sentinel to prevent out of sync with global-refdb
Make sure the the global refdb is never updated to a value that is
different from the local refdb. It may happen that a local operation has
an undetected failure that causes the global-refdb to be updated to a
non-existent value. By checking that the target SHA1 always corresponds
to the local refdb value, any discrepancy would result in an error to
the client and avoid to update the global refdb.

This sentinel should never be needed in Gerrit, however it sometimes
happens that the operation fails at Git level and the ref isn't updated
as expected. Prevent nasty out of sync that would require a Gerrit admin
to action and solve them one-by-one.

Change-Id: If30079abd39f17f3859aa3f8b300b65cc0b7f117