Run validation if only required by one of the enforcement methods
As long as one enforcement method (legacy or new) ignored a ref,
it was ignored. That was the wrong priority. It should rather be
that as soon as at least one enforcement method requires validation
it should be validated.

This is especially true since the requiresAll option did not work,
since the legacy method by default ignores all patchset refs for
example.

Change-Id: I2f9cc529ae28dc61f4d392488aef812a52801702