Handle null putValue in global refdb comparison
Previously, it was assumed that, upon ref deletion, a `refPair.putValue`
would default to the zero object ID when null, but this was not the
case. Even compareAndPut on L244 has to deal with ref.putValue being 0
and transforming it to 0s.

Ensure that the sentinel expands a null refPair.putValue to o the Zero
object ID so that so that it behaves the same on both sides of the
check.

Change-Id: I5514845a778d8d0a83c8a042a24d3f64297e63b9