Also expand ref.putValue in globalrefdb sentinel
The expectation was that ref.putValue would return all 0s rather than
null when deleting objects, however that's not true, even
compareAndPut on L244 has to deal with ref.putValue being 0 and
transforming it to 0s.

Ensure that the sentile expands a null ref.putValue to all 0 so that
it behaves the same on both sides of the check.

Change-Id: I5514845a778d8d0a83c8a042a24d3f64297e63b9