Normalise null refs value to ObjectId.zeros()
Gerrit already detects nulls in RefUpdate and sanitise them
with ObjectId.zeros() for consistency with the rest of the code
and the events listeners. However, some parts of the code has
not been covered and ref updates may come through the global-refdb
layer without sanitisation, with potentially causing false-positives
of out of sync or even worse NPEs.

Apply a defensive coat to the SharedRefDbRefUpdate subclass of
RefUpdate to replace nulls with ObjectId.zeros() so that any
missed sanitisation in Gerrit would be compensated at global-refdb
level.

NOTE: The SharedRefDbBatchRefUpdate already contains the relevant
code for sanitisation and properly generates ObjectId.zeros() instead
of nulls for delete and create refs.

But: Issue 430336832
Change-Id: I272f0637ae223ff6aacdc38af4085855fee25af0