Make RefUpdateValidator gracefully handle ref deletions
The RefUpdateValidator makes sure the the global refdb is never updated
to a value that is different from the local refdb. When doing so, it
compares the value the global-refdb is being updated to, with the local
ref value.

In case of ref deletions however, the `RefUpdate` object might carry a
`null` newObjectId, rather than a properly populated zero-id.

This depends on how Gerrit constructs the `refUpdate` object and
historically has happened that some code path were producing `refUpdate`
objects without explicitly setting the `newObjectId` (see change
I0602aa007e and I4c6710cae91, for example).

Handle newObjectIds values explicitly, to be more resilient to others
(or future) ref deletion path that do not explicitly set the
`newObjectId` upon ref deletions.

Bug: Issue 459508611
Change-Id: I4ff912bca271c639da763e9d5087b36c3db6bc5c